requirements_phase1.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ─── Vision ───────────────────────────────────────────────────────────────────
timm==0.9.12
opencv-python==4.9.0.80
albumentations==1.3.1
Pillow==10.2.0
grad-cam==1.4.8

# ─── Data & ML ────────────────────────────────────────────────────────────────
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
tqdm==4.66.1

# ─── Experiment Tracking ──────────────────────────────────────────────────────
wandb==0.16.3

# ─── API ──────────────────────────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
httpx==0.26.0

# ─── UI ───────────────────────────────────────────────────────────────────────
gradio>=4.44.0

# ─── Utils ────────────────────────────────────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.0
reportlab==4.1.0
